Refactored generation tests
